scripts/requirements.txt
scikit-learn
pandas
joblib
psycopg2-binary 